platform: Remove NXP_CHIP_FW_TYPE flag
This flag is used only to pass through this barrier[1]. After
integrating Android.bp into NXP project[2], we can abandon it
and explicitly include the firmware through PRODUCT_PACKAGES.
This change will allow us to get rid of it and eliminate the
extra inclusion of nxp-vendor.mk in device.mk.
The correct version of the firmware was also added, as the
platform comes with the SN100U chip.

[1] https://github.com/sonyxperiadev/vendor-nxp/blob/master/nxp-vendor.mk
[2] https://github.com/sonyxperiadev/vendor-nxp/pull/7

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>