etc: Move NXP NFC configs to platform
All devices on the platform are released with the same NFC
chip and have the same configs. Move libnfc-nxp.conf from
the device repo to the platform, also add the necessary
libnfc-nxp_RF.conf which contains the radio frequency
configuration, and also update libnfc-nci.conf.

In sync with 64.2.A.2.224 stock firmware.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>